+++ PotThrottle.h
@@ -52,8 +52,8 @@ public:
 	void handleTick();
 	void setup();
 	ThrottleStatus getStatus();
-	int getRawThrottle1();
-	int getRawThrottle2();
+	uint16_t getRawThrottle1();
+	uint16_t getRawThrottle2();
 	void saveEEPROM();
 	void saveConfiguration();
